I need your help
Hello there, I 've come across one of your answers on stackoverflow (https://stackoverflow.com/questions/69258754/is-it-possible-to-play-the-same-audio-on-two-audio-output-devices-simultaneously), and I need your help regarding this.

I am finishing my final year project, which is a "Live performance mixer" built with the web audio api.

What I was wondering about your post is, if you do connect the audio element to the audio output device you want, will you be able to still modify the audio with the modification nodes? 

It just like, it feels to me that you just connect the audio to output and that's that, no in between nodes so no modification can be made, am I right?